+++ CHANGELOG.md
@@ -24,9 +24,11 @@
 ---
 
 ## 2026-08/31-MO
-- `CHANGE`: Refactored `scripts/build.sh` to accept `--clear` and `--clear-only` flags
+- `scripts`
+  - `CHANGE:` Refactored `scripts/build.sh` to accept `--clear` and `--clear-only` flags
+  - `NEW:` Added `scripts/add_license_headers.sh` to add/remove predefined copyright and license header.
 - More documentation polish; fleshing out some more key tables in main `README`.
-- `DOCS`: Fixed image paths in `docs/DESIGN.md` and `docs/LEADER.md`.
+- `DOCS:` Fixed image paths in `docs/DESIGN.md` and `docs/LEADER.md`.
 
 ## 2026-08/30-SU
 - **Doom** (1993)